requirements.txt
uvicorn
fastapi
psycopg2-binary
tzdata
pydantic-settings
python-dotenv
